+++ src/layers/poetry.rs
@@ -1,6 +1,6 @@
 use crate::packaging_tool_versions::POETRY_VERSION;
 use crate::python_version::PythonVersion;
-use crate::utils::StreamedCommandError;
+use crate::utils::{FindBundledPipError, StreamedCommandError};
 use crate::{BuildpackError, PythonBuildpack, utils};
 use libcnb::Env;
 use libcnb::build::BuildContext;
@@ -11,7 +11,6 @@ use libcnb::layer::{
 use libcnb::layer_env::{LayerEnv, ModificationBehavior, Scope};
 use libherokubuildpack::log::log_info;
 use serde::{Deserialize, Serialize};
-use std::io;
 use std::path::Path;
 use std::process::Command;
 
@@ -134,7 +133,7 @@ struct PoetryLayerMetadata {
 #[derive(Debug)]
 pub(crate) enum PoetryLayerError {
     InstallPoetryCommand(StreamedCommandError),
-    LocateBundledPip(io::Error),
+    LocateBundledPip(FindBundledPipError),
 }
 
 impl From<PoetryLayerError> for libcnb::Error<BuildpackError> {